tools/requirements.txt
argparse >= 1.4.0, < 2,
python-dotenv >= 1.1.0, < 2,
requests >= 2.32.4, < 3,
requests-toolbelt >= 1.0.0, < 2,
tqdm >= 4.67.1, < 5,
